% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean_sales}
\alias{clean_sales}
\title{Clean the vgsales dataset}
\usage{
clean_sales(vgsales)
}
\arguments{
\item{vgales}{tibble/df; the `vgsales` dataset}
}
\value{
tibble
}
\description{
Removes entries with a `NA` year value
}
