% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{summarise_vgsales}
\alias{summarise_vgsales}
\title{Summarise the `vgsales` dataset}
\usage{
summarise_vgsales(vgsales, lump_publishers = TRUE, lump_n = 5)
}
\arguments{
\item{lump_publishers}{logical; whether to lump publishers not in the top `n` together to an "Other" group. Default is TRUE.}

\item{lump_n}{numeric; the number of top publishers to keep before lumping the others together}

\item{vgales}{tibble/df; the `vgsales` dataset}
}
\value{
tibble
}
\description{
Summarises the dataset by the Publisher, Year, and Country to return the total number of sales in the
regions specified.
}
