% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy_sales}
\alias{tidy_sales}
\title{Tidy the `vgsales` data}
\usage{
tidy_sales(vgsales)
}
\arguments{
\item{vgales}{tibble/df; the `vgsales` dataset}
}
\value{
tibble
}
\description{
Pivots the ..._Sales columns
}
