% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vgsales}
\alias{vgsales}
\title{Video game sales between 1980 and 2020}
\format{
A tibble with 16598 rows and 11 variables:
\itemize{
   \item Rank All-time rank of the global sales
   \item Name Name of the video game
   \item Platform The platform/console the game was primarily released on
   \item Year The year of release
   \item Genre The type/genre of the game
   \item Publisher The game's publisher
   \item NA_Sales Sales (in millions) in North America
   \item EU_Sales Sales (in millions) in Europe
   \item JP_Sales Sales (in millions) in Japan
   \item Other_Sales Sales (in millions) in other regions
   \item Global_Sales Sales (in millions) globally
}
}
\usage{
data(vgsales)
}
\description{
A dataset of video game sales in different regions scraped from [vgchartz.com](vgchartz.com).
}
\keyword{datasets}
